# api-swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/page/get-page-by-id/{pageId}": {
      "get": {
        "operationId": "getPageById",
        "summary": "",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            }
          }
        },
        "tags": ["page"]
      }
    },
    "/api/page/get-page-by-url": {
      "get": {
        "operationId": "getPageByUrl",
        "summary": "",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            }
          }
        },
        "tags": ["page"]
      }
    },
...
  },
  "info": {
    "title": "PubKey API",
    "description": "The Social Solana Explorer",
    "version": "dev",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://pubkey.ngrok.io"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__session"
      }
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": ["Admin", "User"]
      },
      "Identity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "provider": {
            "type": "object"
          },
          "providerId": {
            "type": "string"
          },
          "verified": {
            "type": "boolean"
          }
        },
        "required": ["id", "provider", "providerId", "verified"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "metaUrl": {
            "type": "string"
          },
          "profileUrl": {
            "type": "string"
          },
          "publicKey": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "identities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Identity"
            }
          },
          "followersCount": {
            "type": "integer"
          },
          "followingCount": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "username",
          "bio",
          "avatarUrl",
          "metaUrl",
          "profileUrl",
          "publicKey",
          "role",
          "identities",
          "followersCount",
          "followingCount"
        ]
      },
...
      "PageStatus": {
        "type": "string",
        "enum": ["Draft", "Published"]
      },
      "PageType": {
        "type": "string",
        "enum": ["Degen", "Gaming", "Personal", "Professional"]
      },
      "PageBlockType": {
        "type": "string",
        "enum": ["Header", "Link"]
      },
      "PageBlock": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "data": {
            "type": "object",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/PageBlockType"
          }
        },
        "required": ["id", "name", "order", "type"]
      },
      "Page": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/PageStatus"
          },
          "type": {
            "$ref": "#/components/schemas/PageType"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "viewUrl": {
            "type": "string",
            "nullable": true
          },
          "siteUrl": {
            "type": "string",
            "nullable": true
          },
          "urls": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "owner": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "blocks": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PageBlock"
            }
          }
        },
        "required": ["status", "type"]
      },
...
    }
  }
}